+++ src/main/java/com/fst/sir/bean/ProduitPanierItem.java
@@ -8,6 +8,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Getter
@@ -20,7 +21,7 @@ public class ProduitPanierItem {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private Double prix;
+    private BigDecimal prix;
 
     @OneToOne
     private ProduitBio produitBio;
